feat: split childProcess integration into childProcess and worker integrations
The childProcess integration handled both child process and worker
thread events under one integration. This splits it into two:
- childProcessIntegration: handles child process events only
- workerIntegration: handles worker thread events only (new)

The old childProcessIntegration's captureWorkerErrors option is
deprecated. Users should use workerIntegration() instead.

Fixes #18698
