+++ dev-packages/node-integration-tests/suites/child-process/test.ts
@@ -10,7 +10,7 @@ const WORKER_EVENT: Event = {
         type: 'Error',
         value: 'Test error',
         mechanism: {
-          type: 'auto.child_process.worker_thread',
+          type: 'auto.worker_thread',
           handled: false,
           data: {
             threadId: expect.any(String),
